feat: add sceleton for main table
