+++ src/components/BodyTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './BodyTable'
