+++ src/pages/OrdersPage/OrdersPage.tsx
@@ -6,8 +6,12 @@ import OrdersBelt from '../../components/OrdersBelt'
 import AddEditElementsModal from '../../ui/AddEditElementsModal'
 import DeletedModal from '../../ui/DeletedModal'
 import Sidebar from '../../ui/Sidebar'
+import { useAppState } from '../../context/AppState'
 
 export default function OrdersPage(): JSX.Element {
+  const {
+    state: { openModalAddEditElements },
+  } = useAppState()
   return (
     <>
       <CssBaseline />
@@ -33,7 +37,7 @@ export default function OrdersPage(): JSX.Element {
             gap: 1,
           }}
         >
-          <AddEditElementsModal />
+          {openModalAddEditElements.modal && <AddEditElementsModal />}
           <DeletedModal />
 
           <OrdersBelt />
