+++ src/ui/NoData/NoData.tsx
@@ -4,11 +4,15 @@ import { Stack } from '@mui/joy'
 
 export default function NoData(): JSX.Element {
   return (
-    <Stack display='row' alignItems='center'>
-      <DraftsTwoToneIcon color='primary' sx={{ fontSize: 40 }} />
-      <Typography variant='overline' color='textSecondary'>
-        No Data
-      </Typography>
-    </Stack>
+    <tr>
+      <td colSpan={6} style={{ textAlign: 'center', padding: '20px 0' }}>
+        <Stack display='row' alignItems='center'>
+          <DraftsTwoToneIcon color='primary' sx={{ fontSize: 40 }} />
+          <Typography variant='overline' color='textSecondary'>
+            No Data
+          </Typography>
+        </Stack>
+      </td>
+    </tr>
   )
 }
