+++ src/ui/SkeletonHeadTable/index.ts
@@ -1,0 +1,1 @@
+export { default } from './SkeletonHeadTable'
